decompile battle_intro.s

+++ include/battle_anim_80CA710.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_BATTLE_ANIM_80CA710_H
 #define GUARD_BATTLE_ANIM_80CA710_H
 
-void sub_80E4EF8(int, int, int, int, u16, u8, int);
+void sub_80E4EF8(u8, u8, u8, u8, u16, u8, u8);
 s16 sub_81174E0(s16 a);
 s16 sub_81174C4(s16 a, s16 b);
